Improved dropdowns for pages
